# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CTWall Backend API
  version: 1.0.0
  description: |
    API for ChainThreatWall (CTWall) - SBOM Management and Security Analysis System.
    Adheres to OWASP ASVS Level 2 Security Requirements.
    Uses RFC 7807 for error handling.
servers:
  - url: https://api.ctwall.internal/api/v1
    description: Production Server
  - url: http://localhost:8080/api/v1
    description: Local Development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        API token for user/service accounts/CLI.
        Use: `Authorization: Bearer <API_TOKEN>`.
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Host-CTWALL_SESSION
      description: |
        Session JWT stored in HttpOnly cookie for GUI users.
    RefreshCookieAuth:
      type: apiKey
      in: cookie
      name: __Secure-CTWALL_REFRESH
      description: |
        Refresh token stored in HttpOnly cookie for GUI session renewal.
  schemas:
...
    Error:
      type: object
      description: RFC 7807 Problem Details
      properties:
        type:
          type: string
          maxLength: 120
          format: uri
          example: "about:blank"
        title:
          type: string
          maxLength: 120
          example: "Invalid Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          maxLength: 120
          example: "Field 'sbom_file' is missing."
        instance:
          type: string
          maxLength: 120
          example: "/api/v1/ingest"
...
    DepAlertRevisionVerdict:
      type: object
      required:
        - projectId
        - testId
        - revisionId
        - verdict
        - totalComponents
        - malwareComponents
        - cleanComponents
        - unknownComponents
        - computedAt
      properties:
        projectId:
          type: string
          maxLength: 120
          format: uuid
        testId:
          type: string
          maxLength: 120
          format: uuid
        revisionId:
          type: string
          maxLength: 120
          format: uuid
        verdict:
          type: string
          maxLength: 120
          enum: [MALWARE, CLEAN, UNKNOWN]
        totalComponents:
          type: integer
          minimum: 0
        malwareComponents:
          type: integer
          minimum: 0
        cleanComponents:
          type: integer
          minimum: 0
        unknownComponents:
          type: integer
          minimum: 0
        computedAt:
          type: string
          format: date-time
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
          maxLength: 120
          format: uuid
        actorId:
          type: string
          maxLength: 120
          format: uuid
        action:
          type: string
          maxLength: 200
        entityType:
          type: string
          maxLength: 200
        entityId:
          type: string
          maxLength: 120
          format: uuid
        details:
          type: object
        ipAddress:
          type: string
          maxLength: 200
        createdAt:
          type: string
          maxLength: 120
          format: date-time
    EventAggregate:
      type: object
      properties:
        eventKey:
          type: string
          maxLength: 240
        category:
          type: string
          maxLength: 120
        severity:
          type: string
          maxLength: 16
          enum: [INFO, WARN, ERROR]
        minRole:
          type: string
          maxLength: 16
          enum: [read, write, admin]
        title:
          type: string
          maxLength: 240
        message:
          type: string
          maxLength: 2000
        component:
          type: string
          maxLength: 240
        errorId:
          type: string
          maxLength: 120
        projectId:
          type: string
          maxLength: 120
          format: uuid
        firstSeenAt:
          type: string
          maxLength: 120
          format: date-time
        lastSeenAt:
          type: string
          maxLength: 120
          format: date-time
        occurrences:
          type: integer
        acknowledgedAt:
          type: string
          maxLength: 120
          format: date-time
        status:
          type: string
          maxLength: 32
          enum: [open, acknowledged]
...
    EventsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventAggregate"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    EventDetailsResponse:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/EventAggregate"
        occurrences:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
    DashboardTopItem:
      type: object
      properties:
        name:
          type: string
          maxLength: 240
        value:
          type: integer
    DashboardPackageType:
      type: object
      properties:
        packageType:
          type: string
          maxLength: 120
        count:
          type: integer
        percentage:
          type: number
          format: float
    DashboardMalwarePackage:
      type: object
      properties:
        purl:
          type: string
          maxLength: 2048
        occurrences:
          type: integer
        lastSeenAt:
          type: string
          maxLength: 120
          format: date-time
          nullable: true
    DashboardIngestActivity:
      type: object
      properties:
        id:
          type: string
          maxLength: 120
          format: uuid
        timestamp:
          type: string
          maxLength: 120
          format: date-time
        status:
          type: string
          maxLength: 120
        stage:
          type: string
          maxLength: 120
        componentsImported:
          type: integer
        productId:
          type: string
          maxLength: 120
          format: uuid
          nullable: true
        scopeId:
          type: string
          maxLength: 120
          format: uuid
          nullable: true
        testId:
          type: string
          maxLength: 120
          format: uuid
          nullable: true
        productName:
          type: string
          maxLength: 240
        scopeName:
          type: string
          maxLength: 240
        testName:
          type: string
          maxLength: 240
        errorMessage:
          type: string
          maxLength: 2000
    DashboardIngestBucket:
      type: object
      properties:
        bucketStart:
          type: string
          maxLength: 120
          format: date-time
        imports:
          type: integer
        componentsImported:
          type: integer
    DashboardVerdictDistribution:
      type: object
      properties:
        malware:
          type: integer
        clean:
          type: integer
        unknown:
          type: integer
    DashboardKpis:
      type: object
      properties:
        malwareActive:
          type: integer
        affectedTests:
          type: integer
        oldestPackageScanAt:
          type: string
          maxLength: 120
          format: date-time
          nullable: true
        lastAnalysisAt:
          type: string
          maxLength: 120
          format: date-time
          nullable: true
        lastMalwareSummaryComputedAt:
          type: string
          maxLength: 120
          format: date-time
          nullable: true
        ingestImports:
          type: integer
        ingestComponents:
          type: integer
        overridesActive:
          type: integer
    DashboardOverviewData:
      type: object
      properties:
        generatedAt:
          type: string
          maxLength: 120
          format: date-time
        rangeStart:
          type: string
          maxLength: 120
          format: date-time
        rangeEnd:
          type: string
          maxLength: 120
          format: date-time
        projectCount:
          type: integer
        topN:
          type: integer
        kpis:
          $ref: "#/components/schemas/DashboardKpis"
        verdictDistribution:
          $ref: "#/components/schemas/DashboardVerdictDistribution"
        topProducts:
          type: array
          items:
            $ref: "#/components/schemas/DashboardTopItem"
        topScopes:
          type: array
          items:
            $ref: "#/components/schemas/DashboardTopItem"
        topPackageTypes:
          type: array
          items:
            $ref: "#/components/schemas/DashboardPackageType"
        topMalwarePackages:
          type: array
          items:
            $ref: "#/components/schemas/DashboardMalwarePackage"
        recentIngest:
          type: array
          items:
            $ref: "#/components/schemas/DashboardIngestActivity"
        ingestTrend:
          type: array
          items:
            $ref: "#/components/schemas/DashboardIngestBucket"
    DashboardOverviewResponse:
      type: object
      properties:
        scope:
          type: string
          maxLength: 32
          enum: [project]
        projectId:
          type: string
          maxLength: 120
          format: uuid
        range:
          type: string
          maxLength: 8
          enum: [24h, 7d, 30d, 90d]
        data:
          $ref: "#/components/schemas/DashboardOverviewData"
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /dashboard/overview:
    get:
      summary: Project dashboard overview (active selected project)
      operationId: getDashboardOverview
      parameters:
        - name: range
          in: query
          required: false
          schema:
            type: string
            maxLength: 8
            enum: [24h, 7d, 30d, 90d]
            default: 7d
        - name: topN
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
      responses:
        '200':
          description: Dashboard overview loaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverviewResponse"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: Failed to load dashboard overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /events:
    get:
      summary: List aggregated Events (filtered + paginated, backend-side)
      operationId: listEvents
      parameters:
        - name: severity
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
          description: Comma-separated list (INFO,WARN,ERROR). Default ERROR.
        - name: category_in
          in: query
          required: false
          schema:
            type: string
            maxLength: 400
          description: Legacy filter. Comma-separated list of categories to include (applies to all severities when severity-specific filters are not provided).
        - name: category_in_error
          in: query
          required: false
          schema:
            type: string
            maxLength: 400
          description: Comma-separated list of categories to include for ERROR (optional; empty/omitted = all categories).
        - name: category_in_warn
          in: query
          required: false
          schema:
            type: string
            maxLength: 400
          description: Comma-separated list of categories to include for WARN (optional; empty/omitted = all categories).
        - name: category_in_info
          in: query
          required: false
          schema:
            type: string
            maxLength: 400
          description: Comma-separated list of categories to include for INFO (optional; empty/omitted = all categories).
        - name: status
          in: query
          required: false
          schema:
            type: string
            maxLength: 32
            enum: [open, acknowledged]
          description: Default open.
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 128
          description: Optional search string (title/message/event_key).
        - name: from
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
            format: date-time
          description: Optional RFC3339 start timestamp.
        - name: to
          in: query
          required: false
          schema:
            type: string
            maxLength: 120
            format: date-time
          description: Optional RFC3339 end timestamp.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Events list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventsListResponse"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /events/{eventKey}:
    get:
      summary: Get single event aggregate + recent occurrences
      operationId: getEvent
      parameters:
        - in: path
          name: eventKey
          required: true
          schema:
            type: string
            maxLength: 240
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetailsResponse"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /tests/{testId}/revisions/{revisionId}/depalert-verdict:
    get:
      summary: Get revision malware verdict for DepAlert
      operationId: getDepAlertRevisionVerdict
      parameters:
        - name: testId
          in: path
          required: true
          schema:
            type: string
            maxLength: 120
            format: uuid
        - name: revisionId
          in: path
          required: true
          schema:
            type: string
            maxLength: 120
            format: uuid
      responses:
        '200':
          description: Revision malware verdict for the selected project scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepAlertRevisionVerdict'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Revision not found in selected project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...